Move get_handle_at_point to aspect
The only place where it's used.
