+++ gaphas/aspect/handlefinder.py
@@ -1,5 +1,6 @@
 from functools import singledispatch
 
+from gaphas.aspect.finder import handle_at_point
 from gaphas.item import Item
 from gaphas.types import Pos
 from gaphas.view import GtkView
@@ -13,7 +14,7 @@ class ItemHandleFinder:
         self.view = view
 
     def get_handle_at_point(self, pos: Pos):
-        return self.view.get_handle_at_point(pos)
+        return handle_at_point(self.view, pos)
 
 
 HandleFinder = singledispatch(ItemHandleFinder)
